patch 0.1.0118: Adjust AVR decoder to apply masks.
ticket: https://trac.mcusim.org/ticket/28

Decoder of the AVR instructions was adjusted to apply access masks of
the I/O registers in this patch. It is necessary to prevent firmware
from overwriting bits which can be read only by design.

A good example is SPMCR register of ATmega8A:

        7     6     5   4      3      2     1     0
        SPMIE RWWSB N/A RWWSRE BLBSET PGWRT PGERS SPMEN
        R/W   R     -   R/W    R/W    R/W   R/W   R/W

It contains one read-only bit (6) and one bit which is not used. Writing
these two bits from firmware will not change their values.

It is also appeared that io_regs.h does not contain 'TCCR0' register which
can be found in ATmega8A microcontroller. Its mask and reset value were
also missing.

+++ src/avr/avr_m8a.c
@@ -796,7 +796,7 @@ static void usart_transmit(struct MSIM_AVR *mcu)
 	if ((err == 0) && (IS_CLEAR(DM(UCSRA), UDRE) == 1)) {
 		if (mcu->pty.master_fd >= 0) {
 			written = MSIM_AVR_PTYWrite(mcu, buf, buf_len);
-			if (written != buf_len) {
+			if (written != (int)buf_len) {
 				snprintf(mcu->log, sizeof mcu->log, "failed "
 				         "to feed PTY master with USART data, "
 				         "slave=%s", mcu->pty.slave_name);
@@ -873,7 +873,7 @@ static void usart_receive(struct MSIM_AVR *mcu)
 	if ((err == 0) && (IS_CLEAR(DM(UCSRA), RXC) == 1)) {
 		if (mcu->pty.master_fd >= 0) {
 			recv = MSIM_AVR_PTYRead(mcu, buf, buf_len);
-			if (recv != buf_len) {
+			if (recv != (int)buf_len) {
 				snprintf(mcu->log, sizeof mcu->log, "failed "
 				         "to read USART data from PTY master, "
 				         "slave=%s", mcu->pty.slave_name);
